+++ labs/lab4/mouse.h
@@ -4,7 +4,7 @@
 
 void mouse_int_handler();
 
-int mouse_subscribe_int(uint8_t *bit_no);
+int mouse_subscribe_int(uint16_t *bit_no);
 
 int mouse_unsubscribe_int();
